+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `pileItemBrightness` to control the brightness of pile items
 - Add `pileItemTint` to control the [tint of pile items](https://pixijs.download/dev/docs/PIXI.Sprite.html#tint)
 - Add `randomOffsetRange` and `randomRotationRange` properties
+- Rename `itemOpacity` to `pileItemOpacity` for consistency
 - Change pile border to be scale invariant
 - Animate item rotation
 - Unmagnify magnified piles on drag start